+++ src/cron.h
@@ -1,6 +1,6 @@
 /*
 	dtee - run a program with standard output and standard error copied to files
-	Copyright 2018  Simon Arlott
+	Copyright 2018,2021  Simon Arlott
 
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU General Public License as published by
@@ -41,6 +41,7 @@ public:
 
 	void terminated(int status, int signum, bool core_dump) override;
 	void interrupted(int signum) override;
+	void error(ErrorType type) override;
 
 	bool report();